requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
django==4.1.6
django-templated-mail==1.1.1
djangorestframework==3.14.0
djongo==1.3.6
et-xmlfile==1.1.0 ; python_version >= '3.6'
openpyxl==3.1.0
pillow==9.4.0 ; python_version >= '3.7'
pymongo==3.12.1
pytz==2022.7.1
reportlab==3.6.12
sqlparse==0.2.4
